+++ tests/probly/visualization/credal/test_2d.py
