+++ tests/probly/visualization/credal/test_3d.py
