+++ tests/probly/visualization/credal/test_visualization_integration.py
