+++ tests/probly/visualization/dirichlet/test_dirichlet_plot.py
